pyproject.toml
[project]
name = "optimize-me"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "codeflash>=0.17.3",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "opentelemetry-api>=1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
    "opentelemetry-instrumentation>=0.59b0",
    "opentelemetry-sdk>=1.38.0",
    "pandas>=2.3.3",
]

[tool.codeflash]
# All paths are relative to this pyproject.toml's directory.
module-root = "src"
tests-root = "tests"
test-framework = "pytest"
ignore-paths = []
formatter-cmds = ["black $file"]
